Rebind configuration properties on startup in case environment changed
The main driver for this was a config server use case, where the server
wants to configure its own properties in the remote repo. It was working
up to a point, but the beans in bootstrap context were not refreshed
after the Environment was built so you had to manually refresh them to
get them to point to (e.g.) config repos configured remotely.

Fixes https://github.com/spring-cloud/spring-cloud-config/issues/139