Take care with order of property sources in /refresh endpoint
It's kind of a corner case, but not a very small corner: if a /refresh
leads to a change in a key that was overridden from its default value (e.g.
n a profile-specific config file) then the key is not identified as changed.
The change is still applied.

This commit fixes the key computation and adds a couple of tests.

Fixes gh-73