Change in contract for @LoadBalanced
Users now need to create their own bean and qualify it, and Spring
Cloud will customize it (instead of providing the @Bean itself).

This is much better for users, since they remain in control of the
bean declarations, and can choose which one (if any) is @Primary.
It's a breaking change for some apps (if they rely on a
@LoadBalanced RestTemplate being automatically injected).