GenericScope can delegate actual bean creation to lower level
GenericScope has some requirements to synchronize the bean creation
and destruction phases, which exposes it to potential deadlocks (e.g.
as reported in #383). The fact that the bean factory is also taking
locks is taken advantage of to ensure that the same lock is taken on
bean creation in all cases. So all threads have to wait for the same
lock, and there is no danger of interleaving.