Spring Cloud Config client don't support removing PropertySource
We saw a regression in the Spring Cloud Config client.
We use it with Spring Cloud Config Server servicing files hosted in a
Git repository. We sometimes need to remove some configuration file from
the repository.
Since https://github.com/spring-cloud/spring-cloud-commons/issues/611
was merged, our applications can not detect when such a file is deleted.

The explanation is that before #611, the PropertySources (PS) were
wrapped in a CompositePropertySource, and then ContextRefresher.addConfigFilesToEnvironment() would
match the names and replace the old CompositePropertySource wrapping all
the PS including the one that was removed with the new CompositePropertySource that do not contains it, and all is fine.

After #611, we have a list of PS, and we would only update the existing
PS or add new ones but never detect that a PS was removed.

Here is a proposed fix, but any suggestion or modifications are welcome.