Clarify what is doing the overriding of remote settings
If I am reading this correctly, the existing two lines flip the perspective of what is doing the overriding from “subject/actor is remote (e.g., config-service) overriding local settings” (in the property names) to “subject/actor is local settings overriding remote ones” (in the text) within each line:

Existing:
```
* `spring.cloud.config.overrideNone=true`: Override from any local property source.
* `spring.cloud.config.overrideSystemProperties=false`: Only system properties, command line arguments, and environment variables (but not the local config files) should override the remote settings.
```

My change aims to clarify by not re-using the verb "override", which is in the property names. It also removes "should", which I find puzzling.