Fix DelegatingServiceInstanceListSupplier must
respect SelectedInstanceCallback on its delegate. Motivation behind is
that some load balancer implementations like the RoundRobinLoadBalancer
to call selectedServiceInstance on the ServiceInstanceListSupplier after
choosing the next instance. However, when a ServiceInstanceListSupplier
in wrapped by the DelegatingServiceInstanceListSupplier this
functionality breaks, as DelegatingServiceInstanceListSupplier is not
implementing SelectedInstanceCallback and therfore does ignoring that its
delegate might also do. This might break the functionality of the
delegate. Fixes gh-1221.