Adapt to restructuring changes in Spring Boot 4.0
This commit updates imports and dependencies to adapt to Spring Boot
4.0's restructuring.

Both SimpleDiscoveryClientAutoConfiguration and
SimpleDiscoveryClientAutoConfiguration have been reworked as
ServerProperties and WebServerInitializedEvent may not be on the
classpath due them moving to spring-boot-web-server.

This commit also updates code that was using APIs that were
deprecated in Spring Boot 3.x and that have now been removed.

Signed-off-by: Andy Wilkinson <andy.wilkinson@broadcom.com>